Ancillaries: Added relay to all components.
