+++ app/ancillaryFactory/src/components/NoAncillaryMessage.js
@@ -2,8 +2,7 @@
 
 import * as React from 'react';
 import { View } from 'react-native';
-import { Text } from '@kiwicom/mobile-shared';
-import { Translation } from '@kiwicom/mobile-localization';
+import { Text, Translation } from '@kiwicom/mobile-shared';
 
 const NoAncillary = () => (
   <View>
