+++ app/fastTrack/index.js
@@ -1,3 +1,2 @@
 // @flow strict
-
 export { default as FastTrackBanner } from './src/appRegistry/FastTrackBanner';
