+++ app/fastTrack/src/appRegistry/FastTrackBanner.js
@@ -1,92 +1,40 @@
 // @flow strict
 
 import * as React from 'react';
-import { View } from 'react-native';
-import { Text, TextButton, StyleSheet, Icon } from '@kiwicom/mobile-shared';
-import { Translation } from '@kiwicom/mobile-localization';
-import { defaultTokens } from '@kiwicom/mobile-orbit';
+import { PrivateApiRenderer, graphql } from '@kiwicom/mobile-relay';
 
-import FastTrackModal from '../components/FastTrackModal';
-import Services from '../components/Services';
+import FastTrackBannerContent from '../components/FastTrackBannerContent';
+import type { FastTrackBannerQueryResponse } from './__generated__/FastTrackBannerQuery.graphql';
 
 type Props = {|
   +bookingId: number,
-  +requester: (url: string) => Promise<Response>,
 |};
 
-type State = {|
-  +isModalVisible: boolean,
-  +documentUrl: string,
-|};
+const query = graphql`
+  query FastTrackBannerQuery($bookingId: Int!) {
+    bookingAncillaries(bookingId: $bookingId, attachmentsFor: [FAST_TRACK]) {
+      ...FastTrackBannerContent_data
+    }
+  }
+`;
 
-class FastTrackBanner extends React.Component<Props, State> {
-  state = {
-    isModalVisible: false,
-    documentUrl: '',
-  };
-
-  onOpenModal = () => {
-    this.setState({
-      isModalVisible: true,
-    });
-  };
-
-  onCloseModal = () => {
-    this.setState({
-      isModalVisible: false,
-    });
+class FastTrackBanner extends React.Component<Props> {
+  renderFastTrackBannerContent = (props: FastTrackBannerQueryResponse) => {
+    return <FastTrackBannerContent data={props.bookingAncillaries} />;
   };
 
   render() {
     return (
-      <View style={style.container}>
-        <View style={style.header}>
-          <Icon name="deals" color={defaultTokens.colorIconSecondary} />
-          <Text style={style.title}>
-            <Translation id="fast_track.banner.title" />
-          </Text>
-        </View>
-        <Text style={style.description}>
-          <Translation id="fast_track.banner.description" />
-        </Text>
-        <Services />
-        <TextButton
-          type="primary"
-          onPress={this.onOpenModal}
-          title={<Translation id="fast_track.banner.show_qr_button" />}
-        />
-        <FastTrackModal
-          documentUrl={this.state.documentUrl}
-          isVisible={this.state.isModalVisible}
-          onCloseModal={this.onCloseModal}
-        />
-      </View>
+      <PrivateApiRenderer
+        query={query}
+        variables={{
+          bookingId: this.props.bookingId,
+        }}
+        authHeaderKey={'KW-Auth-Token'}
+        render={this.renderFastTrackBannerContent}
+      />
     );
   }
 }
 
 export default FastTrackBanner;
-
-const style = StyleSheet.create({
-  container: {
-    padding: 16,
-  },
-  header: {
-    flexDirection: 'row',
-    justifyContent: 'flex-start',
-  },
-  title: {
-    marginBottom: 16,
-    marginStart: 8,
-    textAlign: 'center',
-    fontSize: 16,
-    fontWeight: 'bold',
-    color: defaultTokens.paletteInkDark,
-  },
-  description: {
-    fontSize: 14,
-    lineHeight: 19,
-    color: defaultTokens.paletteInkNormal,
-    marginBottom: 24,
-  },
-});
